gcp/esame-16-Gennaio-2025/requirements.txt
flask == 2.3.3
flask-restful
WTForms
requests
google-cloud-firestore
google-cloud-pubsub 
gunicorn
numpy
